requirements.txt
# Core
python-dotenv==1.0.0
pydantic==2.5.0

# Data Processing
pandas==2.1.3
numpy==1.26.2
pyspark==3.5.0

# Storage & APIs
boto3==1.29.7
minio==7.2.0
requests==2.31.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# API
flask==3.0.0
flask-cors==4.0.0
flask-restx==1.3.2

# Visualization
plotly==5.18.0

# Workflow
airflow==2.7.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Logging
python-json-logger==2.0.7

# Code Quality
black==23.12.0
flake8==6.1.0
pylint==3.0.3
